reorganize ItemBase.cs
