+++ TreeDim.StackBuilder.Basics/Helpers/IItemListener.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace treeDiM.StackBuilder.Basics
+{
+    public interface IItemListener
+    {
+        void Update(ItemBase itemFrom);
+        void Kill(ItemBase itemFrom);
+    }
+}
